This is the script I used for work
Heres the full script. Shot and everything you need.

Powershell cmdlet: new-ADUser
parameters as follows -name (full persons name) -givenname (the persons first name) -surname (the persons last name) -samAccountName (the username you want to give them) -userPrinicipleName (basically the email address) -pathj (the OU and domain pathway) -accountPassword (this is to have the user add their own password) -enabled (this is just saying if the AD account is turned on or not).